+++ src/main/resources/init_db.sql
@@ -3,6 +3,6 @@ use bookTest;
     CREATE TABLE books (
         id INT NOT NULL AUTO_INCREMENT,
         title VARCHAR(255),
-        price INT,
+        price DECIMAL(10,2),
         PRIMARY KEY (id)
     );
